Add accessor to internal event_base
Most libevent APIs require an event_base to register timers and fd notifiers.
Integrating a evhttp server with the Qt event-loop requires access to the
event_base for the current thread which is hidden in the internal private class.
This patch makes the event_base accessible to applications that require this
type of integration.